refactor: migrate auth components and services to new structure
- Updated router to import new auth procedures from the new path.
- Removed obsolete auth components and services from zap-old.
- Refactored hooks for handling authentication logic.
