+++ core/src/rpc/router.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import { ai } from "@/zap/ai/rpc/procedures";
-import { auth } from "@/zap-old/rpc/procedures/auth.rpc";
+import { auth } from "@/zap/auth/rpc/procedures";
 import { example } from "@/zap-old/rpc/procedures/example.rpc";
 import { feedbacks } from "@/zap-old/rpc/procedures/feedbacks.rpc";
 import { mails } from "@/zap-old/rpc/procedures/mails.rpc";
