--- core/src/zap-old/services/auth/get-user-id.service.ts
@@ -1,14 +1,0 @@
-import "server-only";
-
-import { AuthenticationError } from "@/zap/lib/api/errors";
-import { getUserService } from "@/zap/services/auth/get-user.service";
-
-export async function getUserIdService() {
-  const user = await getUserService();
-
-  if (!user) {
-    throw new AuthenticationError("User not authenticated");
-  }
-
-  return user.id;
-}
