+++ core/src/zap/auth/components/forms/index.ts
@@ -1,0 +1,4 @@
+export * from "@/zap/auth/components/forms/forgot-password-form";
+export * from "@/zap/auth/components/forms/login-form";
+export * from "@/zap/auth/components/forms/register-form";
+export * from "@/zap/auth/components/forms/reset-password-form";
