+++ core/src/zap/auth/components/index.ts
@@ -1,0 +1,3 @@
+export * from "@/zap/auth/components/auth-page";
+export * from "@/zap/auth/components/forms";
+export * from "@/zap/auth/components/social-provider-button";
