+++ core/src/zap/auth/lib/redirects.ts
@@ -1,17 +1,14 @@
 import "server-only";
 
 import { ZAP_DEFAULT_SETTINGS } from "@/zap.config";
-
-import { auth } from "./server";
+import { getAuthData } from "@/zap/auth/services";
 
 export function redirectToLogin() {
   return Response.redirect(ZAP_DEFAULT_SETTINGS.AUTH.LOGIN_URL, 302);
 }
 
-export async function getAuthDataOrRedirectToLogin(headers: Headers) {
-  const result = await auth.api.getSession({
-    headers,
-  });
+export async function getAuthDataOrRedirectToLogin() {
+  const result = await getAuthData();
   if (!result?.session) {
     return redirectToLogin();
   }
